refactor(linter/vue): extract casing helpers into shared utils
